Move writeFile to promised.js, add wrappers in otherCurrency
